+++ frontend/src/components/CalculationsTable/classes/textAreaReader.ts
@@ -1,0 +1,19 @@
+import { EpidTextAreaSplitted, EpidTextArea } from "../localTypes";
+import { TextAreaValidation } from "./textAreaValidation";
+
+export class TextAreaReader {
+  #textAreaSplitted: EpidTextAreaSplitted;
+
+  constructor(textArea: EpidTextArea) {
+    const trimmed = textArea.content.trimEnd();
+    const splitted = trimmed.split(textArea.delimSymbol);
+    const textAreaSplitted = { ...textArea, content: splitted } as EpidTextAreaSplitted;
+
+    const textAreaValidation = new TextAreaValidation();
+    textAreaValidation.validate(textAreaSplitted);
+    this.#textAreaSplitted = textAreaSplitted;
+  }
+  read() {
+    return this.#textAreaSplitted;
+  }
+}
