+++ frontend/src/components/CalculationsTable/utils.ts
@@ -1,676 +1,4 @@
-import {
-  calculationsPrecision,
-  defaultP,
-  RussiaRegionCode,
-  textAreaTitlesAgeEndChecked,
-  textAreaTitlesAllChecked,
-  textAreaTitlesAllUnchecked,
-  textAreaTitlesSexRecognitionChecked,
-  upperYearBound,
-} from "../../constants";
-import {
-  CalculatedNoSexRecognitionTableRow,
-  CalculatedSexRecognitionTableColumnDataIndex,
-  CalculatedSexRecognitionTableRow,
-  CalculatedTableColumnType,
-  CalculatedTableRow,
-  PlotlyInputData,
-  Sex,
-  TableRowFromTextAreas,
-  TableRowFromTextAreasAgeEndChecked,
-  TableRowFromTextAreasAllChecked,
-  TextAreaContentMeta,
-  TextAreaTitle,
-  TextAreaTitleAgeEndChecked,
-  TextAreaTitleAllChecked,
-} from "../../types";
-import { PopulationSingleYear } from "../../utils";
-import quantile from "@stdlib/stats/base/dists/chisquare/quantile";
-import { TextAreaValidation } from "./textAreaValidation";
-import { EpidTextArea, EpidTextAreaSplitted, InputMode } from "./localTypes";
-import { TextAreasValidation } from "./textAreasValidation";
-import { textAreaAgeEnd } from "./localConstants";
-
-class TextAreaReader {
-  #textAreaSplitted: EpidTextAreaSplitted;
-
-  constructor(textArea: EpidTextArea) {
-    const trimmed = textArea.content.trimEnd();
-    const splitted = trimmed.split(textArea.delimSymbol);
-    const textAreaSplitted = { ...textArea, content: splitted } as EpidTextAreaSplitted;
-
-    const textAreaValidation = new TextAreaValidation();
-    textAreaValidation.validate(textAreaSplitted);
-    this.#textAreaSplitted = textAreaSplitted;
-  }
-  read() {
-    return this.#textAreaSplitted;
-  }
-}
-
-class TextAreasReader {
-  #textAreasSplitted: Map<unknown, EpidTextAreaSplitted>;
-
-  constructor(textAreas: Map<unknown, EpidTextArea>, inputMode: InputMode) {
-    const textAreasSplitted = new Map<unknown, EpidTextAreaSplitted>();
-
-    for (const [dataIndex, textArea] of textAreas.entries()) {
-      const textAreaReader = new TextAreaReader(textArea);
-      textAreasSplitted.set(dataIndex, textAreaReader.read());
-    }
-
-    const textAreasValidation = new TextAreasValidation();
-    textAreasValidation.validate(textAreasSplitted, inputMode, false);
-
-    this.#addAgeEnd(textAreasSplitted);
-    this.#textAreasSplitted = textAreasSplitted;
-  }
-
-  #addAgeEnd(textAreasSplitted: Map<unknown, EpidTextAreaSplitted>) {
-    if (!textAreasSplitted.has("ageEnd")) {
-      const newTextAreasSplitted = new Map(textAreasSplitted);
-      const startAges = textAreasSplitted.get("ageStart")!;
-      const endAges: string[] = [];
-
-      for (let i = 1; i < startAges.content.length; ++i) {
-        endAges.push(String(+startAges.content[i] - 1));
-      }
-      endAges.push(String(upperYearBound));
-
-      newTextAreasSplitted.set("ageEnd", {
-        ...textAreaAgeEnd.find((textArea) => textArea.dataIndex === "ageEnd")!,
-        delimSymbol: "\n", // delimSymbol can be any because this textArea doesn't actually exist
-        content: endAges,
-      });
-    }
-  }
-
-  #addMorbidityRussia() {}
-
-  #addMorbidityChosenRegions() {}
-
-  read() {
-    return this.#textAreasSplitted;
-  }
-}
-
-class TextAreasValidationTitlesException extends TextAreasValidationException {
-  constructor() {
-    super("unsupported set of titles has been passed");
-    this.name = "TextAreasValidationTitlesException";
-  }
-}
-
-class TextAreasValidationMorbidityException extends TextAreasValidationException {
-  constructor() {
-    super("could not get morbidity");
-    this.name = "TextAreasValidationMorbidityException";
-  }
-}
-
-class TextAreasValidationReaderException extends TextAreasValidationException {
-  constructor(message: string) {
-    super(message);
-    this.name = "TextAreasValidationReaderException";
-  }
-}
-
-export class EpidCalculator {
-  #textAreas: Map<TextAreaTitle, string[]>;
-  #regionCodes: string[];
-  #population: PopulationSingleYear;
-  #tableRowsFromTextAreas: TableRowFromTextAreas[]; // another way of representing data from text areas
-  #hasSexRecognition: boolean;
-  #calculatedTableRows: CalculatedTableRow[] = [];
-  #lambdaParam: number | null | undefined;
-  #cParam: number | null | undefined;
-
-  // epidCalculator depends on the data received from the text areas
-  // however: "Конечный возраст", "Число заболевших (Россия)", "Число заболевших (выбран. регионы)" cols might not be present
-
-  // "Конечный возраст" is not present when ageEnd was not checked by user
-  // "Число заболевших (Россия)", "Число заболевших (выбран. регионы)" are absent when there is a sex recognition set by user
-
-  // Still, we might need these cols in the future (for example, to show them later in the table or to plot them)
-  // that's why we have to manually add these cols if they haven't existed previously
-  #addLackingColumns() {
-    if (!this.#textAreas.has("Конечный возраст")) {
-      const startAges = this.#textAreas.get("Начальный возраст")!; // #checkTitles should be called beforehand
-      const endAges: string[] = [];
-      for (let i = 1; i < startAges.length; ++i) {
-        endAges.push(String(+startAges[i] - 1));
-      }
-      endAges.push(String(upperYearBound));
-      this.#textAreas.set("Конечный возраст", endAges);
-    }
-
-    if (!this.#textAreas.has("Число заболевших (Россия)")) {
-      const menPopulationRussia = this.#textAreas.get("Число заболевших (мужчины, Россия)")!;
-      const womenPopulationRussia = this.#textAreas.get("Число заболевших (женщины, Россия)")!;
-
-      const populationRussia: string[] = [];
-      for (let i = 0; i < menPopulationRussia.length; ++i) {
-        populationRussia.push(String(+menPopulationRussia[i] + +womenPopulationRussia[i]));
-      }
-      this.#textAreas.set("Число заболевших (Россия)", populationRussia);
-    }
-
-    if (!this.#textAreas.has("Число заболевших (выбран. регионы)")) {
-      const menPopulationChosenRegions = this.#textAreas.get("Число заболевших (мужчины, выбран. регионы)")!;
-      const womenPopulationChosenRegions = this.#textAreas.get("Число заболевших (женщины, выбран. регионы)")!;
-
-      const populationChosenRegions: string[] = [];
-      for (let i = 0; i < menPopulationChosenRegions.length; ++i) {
-        populationChosenRegions.push(String(+menPopulationChosenRegions[i] + +womenPopulationChosenRegions[i]));
-      }
-      this.#textAreas.set("Число заболевших (выбран. регионы)", populationChosenRegions);
-    }
-  }
-
-  constructor(textAreas: EpidTextArea[], population: PopulationSingleYear, regionCodes: string[]) {
-    this.#textAreas = new Map();
-    this.#population = population;
-    this.#regionCodes = regionCodes;
-
-    try {
-      for (const [title, contentMeta] of textAreas.entries()) {
-        const textAreaReader = new TextAreaReader(
-          contentMeta.content,
-          contentMeta.delimSymbol,
-          contentMeta.allowOnlyIntegers,
-          contentMeta.upperBound
-        );
-        this.#textAreas.set(title, textAreaReader.read());
-      }
-    } catch (error) {
-      throw new EpidCalculatorReaderException((error as TextAreaReaderException).message);
-    }
-
-    this.#checkTitles();
-    this.#checkSameLength();
-    this.#checkAges();
-
-    this.#addLackingColumns();
-
-    if (this.#textAreas.has("Число заболевших (женщины, Россия)")) {
-      this.#hasSexRecognition = true;
-    } else {
-      this.#hasSexRecognition = false;
-    }
-
-    this.#tableRowsFromTextAreas = [];
-
-    for (let i = 0; i < this.#textAreas.get("Конечный возраст")!.length; ++i) {
-      const obj = {} as TableRowFromTextAreas;
-
-      if (this.#hasSexRecognition) {
-        for (const key of this.#textAreas.keys() as IterableIterator<TextAreaTitleAllChecked>) {
-          (obj as TableRowFromTextAreasAllChecked)[key] = this.#textAreas.get(key)![i];
-        }
-      } else {
-        for (const key of this.#textAreas.keys() as IterableIterator<TextAreaTitleAgeEndChecked>) {
-          (obj as TableRowFromTextAreasAgeEndChecked)[key] = this.#textAreas.get(key)![i];
-        }
-      }
-      this.#tableRowsFromTextAreas.push(obj);
-    }
-    // console.log(this.#tableRowsFromTextAreas);
-  }
-
-  calculateTable() {
-    if (this.#calculatedTableRows.length) {
-      return this.#calculatedTableRows;
-    }
-    const res = [] as CalculatedTableRow[];
-
-    for (const row of this.#tableRowsFromTextAreas) {
-      const obj = {} as CalculatedTableRow;
-      const k1 = (obj["startAge"] = +row["Начальный возраст"]);
-      const k2 = (obj["endAge"] = +row["Конечный возраст"]);
-
-      // console.log(k1, k2);
-
-      obj["populationRussia"] = this.#population.n(k1, k2);
-
-      obj["morbidityRussia"] = this.getMorbidity(k1, k2);
-      obj["intensiveMorbidityRussia"] = this.getIntensiveMorbidity(k1, k2);
-
-      const intensiveMorbidityRussiaCI = this.getIntensiveMorbidityConfidenceInterval(k1, k2);
-
-      obj["lowerIntensiveMorbidityRussia"] = intensiveMorbidityRussiaCI[0];
-      obj["upperIntensiveMorbidityRussia"] = intensiveMorbidityRussiaCI[1];
-
-      obj["populationChosenRegions"] = this.#population.n(k1, k2, undefined, this.#regionCodes);
-      obj["morbidityChosenRegions"] = this.getMorbidity(k1, k2, undefined, this.#regionCodes);
-      obj["intensiveMorbidityChosenRegions"] = this.getIntensiveMorbidity(k1, k2, undefined, this.#regionCodes);
-
-      const intensiveMorbidityChosenRegionsCI = this.getIntensiveMorbidityConfidenceInterval(
-        k1,
-        k2,
-        undefined,
-        this.#regionCodes
-      );
-
-      obj["lowerIntensiveMorbidityChosenRegions"] = intensiveMorbidityChosenRegionsCI[0];
-      obj["upperIntensiveMorbidityChosenRegions"] = intensiveMorbidityChosenRegionsCI[1];
-
-      if (this.#hasSexRecognition) {
-        (obj as CalculatedSexRecognitionTableRow)["menPopulationRussia"] = this.#population.n(k1, k2, "male");
-        (obj as CalculatedSexRecognitionTableRow)["menMorbidityRussia"] = this.getMorbidity(k1, k2, "male");
-        (obj as CalculatedSexRecognitionTableRow)["menIntensiveMorbidityRussia"] = this.getIntensiveMorbidity(
-          k1,
-          k2,
-          "male"
-        );
-
-        const menIntensiveMorbidityRussiaCI = this.getIntensiveMorbidityConfidenceInterval(k1, k2, "male");
-
-        (obj as CalculatedSexRecognitionTableRow)["menLowerIntensiveMorbidityRussia"] =
-          menIntensiveMorbidityRussiaCI[0];
-        (obj as CalculatedSexRecognitionTableRow)["menUpperIntensiveMorbidityRussia"] =
-          menIntensiveMorbidityRussiaCI[1];
-
-        (obj as CalculatedSexRecognitionTableRow)["womenPopulationRussia"] = this.#population.n(k1, k2, "female");
-        (obj as CalculatedSexRecognitionTableRow)["womenMorbidityRussia"] = this.getMorbidity(k1, k2, "female");
-        (obj as CalculatedSexRecognitionTableRow)["womenIntensiveMorbidityRussia"] = this.getIntensiveMorbidity(
-          k1,
-          k2,
-          "female"
-        );
-
-        const womenIntensiveMorbidityRussiaCI = this.getIntensiveMorbidityConfidenceInterval(k1, k2, "female");
-
-        (obj as CalculatedSexRecognitionTableRow)["womenLowerIntensiveMorbidityRussia"] =
-          womenIntensiveMorbidityRussiaCI[0];
-        (obj as CalculatedSexRecognitionTableRow)["womenUpperIntensiveMorbidityRussia"] =
-          womenIntensiveMorbidityRussiaCI[1];
-
-        (obj as CalculatedSexRecognitionTableRow)["menPopulationChosenRegions"] = this.#population.n(
-          k1,
-          k2,
-          "male",
-          this.#regionCodes
-        );
-        (obj as CalculatedSexRecognitionTableRow)["menMorbidityChosenRegions"] = this.getMorbidity(
-          k1,
-          k2,
-          "male",
-          this.#regionCodes
-        );
-        (obj as CalculatedSexRecognitionTableRow)["menIntensiveMorbidityChosenRegions"] = this.getIntensiveMorbidity(
-          k1,
-          k2,
-          "male",
-          this.#regionCodes
-        );
-
-        const menIntensiveMorbidityChosenRegionsCI = this.getIntensiveMorbidityConfidenceInterval(
-          k1,
-          k2,
-          "male",
-          this.#regionCodes
-        );
-
-        (obj as CalculatedSexRecognitionTableRow)["menLowerIntensiveMorbidityChosenRegions"] =
-          menIntensiveMorbidityChosenRegionsCI[0];
-        (obj as CalculatedSexRecognitionTableRow)["menUpperIntensiveMorbidityChosenRegions"] =
-          menIntensiveMorbidityChosenRegionsCI[1];
-
-        (obj as CalculatedSexRecognitionTableRow)["womenPopulationChosenRegions"] = this.#population.n(
-          k1,
-          k2,
-          "female",
-          this.#regionCodes
-        );
-        (obj as CalculatedSexRecognitionTableRow)["womenMorbidityChosenRegions"] = this.getMorbidity(
-          k1,
-          k2,
-          "female",
-          this.#regionCodes
-        );
-        (obj as CalculatedSexRecognitionTableRow)["womenIntensiveMorbidityChosenRegions"] = this.getIntensiveMorbidity(
-          k1,
-          k2,
-          "female",
-          this.#regionCodes
-        );
-
-        const womenIntensiveMorbidityChosenRegionsCI = this.getIntensiveMorbidityConfidenceInterval(
-          k1,
-          k2,
-          "female",
-          this.#regionCodes
-        );
-
-        (obj as CalculatedSexRecognitionTableRow)["womenLowerIntensiveMorbidityChosenRegions"] =
-          womenIntensiveMorbidityChosenRegionsCI[0];
-        (obj as CalculatedSexRecognitionTableRow)["womenUpperIntensiveMorbidityChosenRegions"] =
-          womenIntensiveMorbidityChosenRegionsCI[1];
-      }
-      res.push(obj);
-    }
-    this.#calculatedTableRows = res;
-    return res;
-  }
-
-  getRussiaMorbidity() {
-    const morbiditiesRussia = this.#textAreas.get("Число заболевших (Россия)")!;
-    return morbiditiesRussia.reduce((sum, curr) => sum + Number(curr), 0);
-  }
-
-  getRussiaIntensiveMorbidity() {
-    const a = this.getRussiaMorbidity();
-    return (10 ** 5 * a) / this.#population.n(0, upperYearBound);
-  }
-
-  getChosenRegionsMorbidity() {
-    const morbiditiesChosenRegions = this.#textAreas.get("Число заболевших (выбран. регионы)")!;
-    return morbiditiesChosenRegions.reduce((sum, curr) => sum + Number(curr), 0);
-  }
-
-  getChosenRegionsIntensiveMorbidity() {
-    const a = this.getChosenRegionsMorbidity();
-    return (10 ** 5 * a) / this.#population.n(0, upperYearBound, undefined, this.#regionCodes);
-  }
-
-  getChosenRegionsStandardizedMorbidity() {
-    let res = 0;
-    if (this.#hasSexRecognition) {
-      for (const row of this.#calculatedTableRows as CalculatedSexRecognitionTableRow[]) {
-        const k1 = row["startAge"];
-        const k2 = row["endAge"];
-
-        res += row["menMorbidityChosenRegions"] * this.#population.h(k1, k2, "male");
-
-        res += row["womenMorbidityChosenRegions"] * this.#population.h(k1, k2, "female");
-      }
-    } else {
-      for (const row of this.#calculatedTableRows as CalculatedNoSexRecognitionTableRow[]) {
-        const k1 = row["startAge"];
-        const k2 = row["endAge"];
-
-        res += row["morbidityChosenRegions"] * this.#population.h(k1, k2);
-      }
-    }
-    return res;
-  }
-
-  getChosenRegionsStandardizedIntensiveMorbidity() {
-    let res = 0;
-    if (this.#hasSexRecognition) {
-      for (const row of this.#calculatedTableRows as CalculatedSexRecognitionTableRow[]) {
-        const k1 = row["startAge"];
-        const k2 = row["endAge"];
-
-        res += row["menIntensiveMorbidityChosenRegions"] * this.#population.h(k1, k2, "male");
-
-        res += row["womenIntensiveMorbidityChosenRegions"] * this.#population.h(k1, k2, "female");
-      }
-    } else {
-      for (const row of this.#calculatedTableRows as CalculatedNoSexRecognitionTableRow[]) {
-        const k1 = row["startAge"];
-        const k2 = row["endAge"];
-
-        res += row["intensiveMorbidityChosenRegions"] * this.#population.h(k1, k2);
-      }
-    }
-    return res;
-  }
-
-  getMorbidity(
-    k1: number,
-    k2: number,
-    m?: Sex,
-    regionCodes?: string[] // if regionCodes are not passed, assume that we take whole Russia
-  ) {
-    const res = this.#tableRowsFromTextAreas.find(
-      (row) => +row["Начальный возраст"] === k1 && +row["Конечный возраст"] === k2
-    );
-    if (!res) {
-      throw new EpidCalculatorMorbidityException();
-    }
-    let a: number;
-    switch (m) {
-      case "male":
-        if (regionCodes) {
-          a = +(res as TableRowFromTextAreasAllChecked)["Число заболевших (мужчины, выбран. регионы)"]!;
-        } else {
-          a = +(res as TableRowFromTextAreasAllChecked)["Число заболевших (мужчины, Россия)"]!;
-        }
-        break;
-      case "female":
-        if (regionCodes) {
-          a = +(res as TableRowFromTextAreasAllChecked)["Число заболевших (женщины, выбран. регионы)"]!;
-        } else {
-          a = +(res as TableRowFromTextAreasAllChecked)["Число заболевших (женщины, Россия)"]!;
-        }
-        break;
-      default:
-        if (regionCodes) {
-          a = +(res as TableRowFromTextAreasAgeEndChecked)["Число заболевших (выбран. регионы)"]!;
-        } else {
-          a = +(res as TableRowFromTextAreasAgeEndChecked)["Число заболевших (Россия)"]!;
-        }
-      // console.log(res);
-      // console.log(a);
-    }
-    return a;
-  }
-
-  getIntensiveMorbidity(
-    k1: number,
-    k2: number,
-    m?: Sex,
-    regionCodes?: string[] // if regionCodes are not passed, assume that we take whole Russia
-  ) {
-    const a = this.getMorbidity(k1, k2, m, regionCodes);
-    const n = this.#population.n(k1, k2, m, regionCodes); // total population in the chosen group
-    return +((10 ** 5 * a) / n).toFixed(calculationsPrecision);
-  }
-
-  // assume morbidity has Poisson distribution
-  #getMorbidityConfidenceInterval(morbidity: number, p = defaultP) {
-    const lowerBound = 0.5 * quantile(p / 2, 2 * morbidity);
-    const upperBound = 0.5 * quantile(1 - p / 2, 2 * (morbidity + 1));
-    return [lowerBound, upperBound];
-  }
-
-  getIntensiveMorbidityConfidenceInterval(
-    k1: number,
-    k2: number,
-    m?: Sex,
-    regionCodes?: string[], // if regionCodes are not passed, assume that we take whole Russia
-    p = defaultP
-  ) {
-    const a = this.getMorbidity(k1, k2, m, regionCodes);
-    const [lowerMorbidity, upperMorbidity] = this.#getMorbidityConfidenceInterval(a, p);
-    const n = this.#population.n(k1, k2, m, regionCodes); // total population in the chosen group
-    return [
-      +((10 ** 5 * lowerMorbidity) / n).toFixed(calculationsPrecision),
-      +((10 ** 5 * upperMorbidity) / n).toFixed(calculationsPrecision),
-    ];
-  }
-
-  #leastSquares1D(x: number[], y: number[], foo: (x: number, param: number) => number, step = 0.01): number {
-    if (x.length !== y.length) {
-      throw new Error("x and y should be of the same length");
-    }
-    let lowestDelta = Infinity;
-    let res = 0;
-    for (let param = step; param < 1; param += step) {
-      const yModel = x.map((curr) => foo(curr, param));
-      const delta = y.map((curr, index) => (curr - yModel[index]) ** 2).reduce((sum, curr) => sum + curr, 0);
-      // console.log(`lambda = ${param}, delta = ${delta}`);
-      if (delta < lowestDelta) {
-        console.log("ymodel =");
-        console.log(yModel);
-        lowestDelta = delta;
-        res = param;
-      }
-    }
-    return res;
-  }
-
-  #findRowWithAgeWithin(age: number) {
-    return this.#calculatedTableRows.find((row) => row.startAge <= age && row.endAge >= age);
-  }
-
-  #relativeMorbidityFunction(age: number, lambda: number) {
-    return lambda * Math.exp(-lambda * age);
-  }
-
-  #morbidityFunction(age: number, lambda: number) {
-    const population = this.#population.n(age);
-    return Math.round(this.#relativeMorbidityFunction(age, lambda) * population);
-  }
-
-  static mapRegionCodes(regionCodes?: string[]) {
-    if (!regionCodes || !regionCodes.length || regionCodes[0] === RussiaRegionCode) {
-      return "Russia";
-    } else {
-      return "ChosenRegions";
-    }
-  }
-
-  static mapSex(sex?: Sex) {
-    if (sex === "male") {
-      return "men";
-    } else if (sex === "female") {
-      return "women";
-    } else {
-      return "";
-    }
-  }
-
-  // keyType = "population" sex = "male", regionCodes is not passed => returns "menPopulationRussia"
-  // keyType = "population" sex = "male", regionCodes is passed => returns "menPopulationChosenRegions"
-  // etc...
-  static getCalculatedTableRowKey(keyType: CalculatedTableColumnType, sex?: Sex, regionCodes?: string[]) {
-    const mappedRegionCodes = this.mapRegionCodes(regionCodes);
-    const mappedSex = this.mapSex(sex);
-    let res = keyType + mappedRegionCodes;
-    if (sex === "male" || sex === "female") {
-      res = mappedSex + res[0].toUpperCase() + res.slice(1);
-    }
-    return res as CalculatedSexRecognitionTableColumnDataIndex;
-  }
-
-  // morbidity data (which is typed by user) may only be available for a range of ages (not for each single age)
-  // however, there might be a necessity to get such data for each single year
-  // idea: having age range [t1, t2] and intensive morbidity "i"
-  // assume that "i_single_year(t)" equals to "i" for each "t" from [t1, t2]
-  getLambdaEstimation(sex?: Sex, regionCodes?: string[]) {
-    if (this.#lambdaParam) {
-      return this.#lambdaParam;
-    }
-    const x = []; // x-coords
-    const y = []; // y-coords
-
-    for (let i = 0; i <= upperYearBound; ++i) {
-      const row = this.#findRowWithAgeWithin(i);
-      // in theory (if everything I've previously written is correct) this thing can only throw error
-      // when a user has chosen to explicitly specify the end ages (thus some ages might be skipped)
-      if (!row) {
-        continue;
-        // throw new Error("smth bad happened...");
-      }
-      x.push(i);
-
-      const population = this.#population.n(i, i, sex, regionCodes);
-      const key = EpidCalculator.getCalculatedTableRowKey("intensiveMorbidity", sex, regionCodes);
-      const intensiveMorbidity = (row as CalculatedSexRecognitionTableRow)[key];
-      console.warn(`key = ${key}, intensiveMorb = ${intensiveMorbidity}`);
-
-      if (intensiveMorbidity === null || intensiveMorbidity === undefined) {
-        throw new Error("wrong getLambdaEstimation usage");
-      }
-
-      y.push(Math.round(intensiveMorbidity / 10 ** 5) * population);
-    }
-
-    console.log(x);
-    console.log(y);
-
-    this.#lambdaParam = this.#leastSquares1D(x, y, this.#morbidityFunction.bind(this));
-    return this.#lambdaParam;
-  }
-
-  getCEstimation(sex?: Sex, regionCodes?: string[]) {
-    const lambda = this.getLambdaEstimation(sex, regionCodes);
-
-    const cValues = [];
-
-    for (let i = 0; i <= upperYearBound; ++i) {
-      const row = this.#findRowWithAgeWithin(i);
-
-      // in theory (if everything I've previously written is correct) this thing can only throw error
-      // when a user has chosen to explicitly specify the end ages (thus some ages might be skipped)
-      if (!row) {
-        continue;
-        // throw new Error("smth bad happened...");
-      }
-      const population = this.#population.n(i, i, sex, regionCodes);
-      const key = EpidCalculator.getCalculatedTableRowKey("intensiveMorbidity", sex, regionCodes);
-      const intensiveMorbidity = (row as CalculatedSexRecognitionTableRow)[key];
-
-      if (intensiveMorbidity === null || intensiveMorbidity === undefined) {
-        throw new Error("wrong getLambdaEstimation usage");
-      }
-
-      const cValue = ((intensiveMorbidity / 10 ** 5) * population) / this.#morbidityFunction(i, lambda);
-      cValues.push(cValue);
-    }
-    console.log(cValues);
-
-    this.#cParam =
-      cValues.filter((cValue) => !isNaN(cValue) && isFinite(cValue)).reduce((sum, curr) => sum + curr) / cValues.length;
-    return this.#cParam;
-  }
-
-  getContactNumber(sex?: Sex, regionCodes?: string[]) {
-    const lambda = this.getLambdaEstimation(sex, regionCodes);
-    const step = 1;
-    let res = 0.1;
-    for (let i = 0; i < upperYearBound; i += step) {
-      const flooredI = Math.floor(i);
-      const row = this.#findRowWithAgeWithin(flooredI);
-      if (!row) {
-        continue;
-      }
-      const h = this.#population.h(flooredI, flooredI, sex, regionCodes);
-      res += this.#relativeMorbidityFunction(i, lambda) * h * step;
-    }
-    return res;
-  }
-
-  getAbsoluteError(sex?: Sex, regionCodes?: string[]) {
-    const lambda = this.getLambdaEstimation(sex, regionCodes);
-    const c = this.getCEstimation(sex, regionCodes);
-    let error = 0;
-
-    for (let i = 0; i <= upperYearBound; ++i) {
-      const row = this.#findRowWithAgeWithin(i);
-      // in theory (if everything I've previously written is correct) this thing can only throw error
-      // when a user has chosen to explicitly specify the end ages (thus some ages might be skipped)
-      if (!row) {
-        continue;
-        // throw new Error("smth bad happened...");
-      }
-      const population = this.#population.n(i, i, sex, regionCodes);
-      const key = EpidCalculator.getCalculatedTableRowKey("intensiveMorbidity", sex, regionCodes);
-      const intensiveMorbidity = (row as CalculatedSexRecognitionTableRow)[key];
-      if (intensiveMorbidity === null || intensiveMorbidity === undefined) {
-        throw new Error("wrong getLambdaEstimation usage");
-      }
-
-      error +=
-        (Math.round(intensiveMorbidity / 10 ** 5) * population -
-          this.#relativeMorbidityFunction(i, lambda) * population * c) **
-        2;
-    }
-    return error;
-  }
-}
+import { CalculatedTableRow, PlotlyInputData } from "../../types";
 
 export function extractDataForPlotting(calculatedTableRows: CalculatedTableRow[], hasSexRecognition: boolean) {
   const res: PlotlyInputData[] = [];
